chore: create src folder and move codes into src folder
