+++ src/models/Chapter.js
@@ -1,4 +1,4 @@
-const mongoose = require("../database");
+const mongoose = require("../../database");
 
 const chapterSchema = new mongoose.Schema(
     {
