+++ src/models/FavoriteManga.js
