+++ src/models/Manga.js
